+++ packages/components/src/DateTimePickers/InputDateRangePicker/InputDateRangePicker.component.js
@@ -2,6 +2,7 @@ import React, { useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import omit from 'lodash/omit';
 import classnames from 'classnames';
+import { Popper } from 'react-popper';
 
 import FocusManager from '../../FocusManager';
 import { focusOnCalendar } from '../../Gesture/withCalendarGesture';
@@ -50,7 +51,74 @@ export default function InputDateRangePicker(props) {
 		handlers.onChange(event, payload, inputRef.current);
 	}
 	return (
-		<div />
+		<DateRange.Manager
+			startDate={props.startDate}
+			endDate={props.endDate}
+			dateFormat={props.dateFormat}
+			onChange={onChange}
+		>
+			<DateRangeContext.Consumer>
+				{({ inputManagement, startDate, endDate }) => {
+					const { onStartChange, onEndChange } = inputManagement;
+					return (
+						<FocusManager
+							className={classnames(theme['date-picker'], 'date-picker')}
+							divRef={containerRef}
+							onClick={handlers.onClick}
+							onFocusIn={handlers.onFocus}
+							onFocusOut={handlers.onBlur}
+							onKeyDown={event => {
+								handlers.onKeyDown(event, inputRef.current);
+							}}
+						>
+							<DateRange.Input
+								{...inputProps}
+								id={`${props.id}-start-input`}
+								date={startDate}
+								onChange={onStartChange}
+								onFocus={() => setInputRef(startDateInputRef)}
+								label={props.t('TC_DATE_PICKER_RANGE_FROM', { defaultValue: 'From' })}
+								ref={startDateInputRef}
+							/>
+							<span className={classnames(theme.arrow, 'arrow')}>
+								<Icon name="talend-arrow-right" className={classnames(theme.icon, 'icon')} />
+							</span>
+							<DateRange.Input
+								{...inputProps}
+								id={`${props.id}-end-input`}
+								date={endDate}
+								onChange={onEndChange}
+								onFocus={() => setInputRef(endDateInputRef)}
+								label={props.t('TC_DATE_PICKER__RANGE_TO', { defaultValue: 'To' })}
+								ref={endDateInputRef}
+							/>
+							{handlers.showPicker && inputRef && (
+								<Popper
+									key="popper"
+									modifiers={{
+										hide: {
+											enabled: false,
+										},
+										preventOverflow: {
+											enabled: false,
+										},
+									}}
+									placement={handlers.getPopperPlacement(inputRef.current)}
+									positionFixed
+									referenceElement={inputRef.current}
+								>
+									{({ ref, style }) => (
+										<div id={popoverId} className={theme.popper} style={style} ref={ref}>
+											<DateRange.Picker {...props} focusedInput={getFocusedInput()} />
+										</div>
+									)}
+								</Popper>
+							)}
+						</FocusManager>
+					);
+				}}
+			</DateRangeContext.Consumer>
+		</DateRange.Manager>
 	);
 }
 InputDateRangePicker.displayName = 'InputDateRangePicker';