+++ packages/components/src/Typeahead/Typeahead.component.renderers.js
@@ -6,8 +6,11 @@ import ControlLabel from 'react-bootstrap/lib/ControlLabel';
 import FormControl from 'react-bootstrap/lib/FormControl';
 import DebounceInput from 'react-debounce-input';
 import classNames from 'classnames';
+import { Popper } from 'react-popper';
 
+import Icon from '../Icon';
 import CircularProgress from '../CircularProgress';
+import Emphasis from '../Emphasis';
 import theme from './Typeahead.scss';
 
 export function renderInputComponent(props) {
@@ -32,7 +35,42 @@ export function renderInputComponent(props) {
 		hasCaret && theme['typeahead-input-caret'],
 	);
 	return (
-		<div />
+		<div className={typeaheadContainerIconClasses}>
+			<ControlLabel srOnly htmlFor={key}>
+				Search
+			</ControlLabel>
+			{debounceMinLength || debounceTimeout ? (
+				<DebounceInput
+					autoFocus
+					id={key}
+					{...rest}
+					disabled={disabled}
+					readOnly={readOnly}
+					debounceTimeout={debounceTimeout}
+					element={FormControl}
+					minLength={debounceMinLength}
+					inputRef={node => {
+						// eslint-disable-next-line react/no-find-dom-node
+						inputRef.current = ReactDOM.findDOMNode(node);
+					}}
+				/>
+			) : (
+				<FormControl
+					id={key}
+					autoFocus
+					{...rest}
+					disabled={disabled}
+					readOnly={readOnly}
+					inputRef={inputRef}
+				/>
+			)}
+			{hasIcon && (
+				<div className={classNames(theme['icon-cls'], hasCaret && theme['icon-caret'])}>
+					{icon && <Icon {...icon} />}
+					{hasCaret && <Icon name="talend-caret-down" />}
+				</div>
+			)}
+		</div>
 	);
 }
 
@@ -49,6 +87,30 @@ renderInputComponent.propTypes = {
 	disabled: PropTypes.bool,
 	readOnly: PropTypes.bool,
 };
+
+function computePopperPosition(data) {
+	const GAP = 45; // the offset between the end of items container and screen boundaries
+	const inputDimensions = data.offsets.reference;
+	const { top, height } = inputDimensions;
+	const offsetTop = top - GAP;
+	const offsetBottom = window.innerHeight - top - height - GAP;
+	const placements = data.placement.split('-');
+	let newPlacement = data.placement;
+	if (placements[0] === 'top' && offsetBottom > offsetTop) {
+		newPlacement = `bottom-${placements[1]}`;
+	}
+	const maxHeight = newPlacement.includes('top') ? offsetTop : offsetBottom;
+
+	return {
+		...data,
+		placement: newPlacement,
+		styles: {
+			...data.styles,
+			width: inputDimensions.width,
+			maxHeight,
+		},
+	};
+}
 
 export function renderItemsContainerFactory(
 	items,
@@ -97,7 +159,62 @@ export function renderItemsContainerFactory(
 		}
 
 		return (
-			<div />
+			<Popper
+				modifiers={{
+					hide: {
+						enabled: false,
+					},
+					preventOverflow: {
+						enabled: false,
+					},
+					shift: {
+						enabled: false,
+					},
+					computePosition: {
+						enabled: true,
+						fn: computePopperPosition,
+					},
+				}}
+				positionFixed
+				boundariesElement="viewport"
+				referenceElement={inputRef.current}
+				placement="bottom-start"
+			>
+				{({ placement = '', ref, scheduleUpdate, style }) => {
+					if (placement.includes('top')) {
+						// @see https://github.com/FezVrasta/react-popper/issues/283#issuecomment-512879262
+						scheduleUpdate();
+					}
+
+					return (
+						<div
+							className={containerClassName}
+							id={containerProps.id}
+							key={containerProps.key}
+							ref={ref}
+							role={containerProps.role}
+							style={style}
+						>
+							<div
+								ref={containerProps.ref}
+								className={theme['items-body']}
+								style={{ maxHeight: style.maxHeight }}
+							>
+								{render(
+									content,
+									{
+										isShown,
+										loading,
+										noResult,
+										searching,
+									},
+									containerProps.ref,
+								)}
+							</div>
+						</div>
+					);
+				}}
+			</Popper>
 		);
 	}
 
@@ -112,7 +229,14 @@ export function renderItemsContainerFactory(
 export function renderSectionTitle(section) {
 	if (section && (section.icon || section.title)) {
 		return (
-			<div />
+			<div className={classNames(theme['section-header'], 'tc-typeahead-section-header')}>
+				{section.icon && <Icon name={section.icon.name} title={section.icon.title} />}
+				<span
+					className={classNames(theme['section-header-title'], 'tc-typeahead-section-header-title')}
+				>
+					{section.title}
+				</span>
+			</div>
 		);
 	}
 	return null;
@@ -128,6 +252,25 @@ export function renderItem(item, { value, ...rest }) {
 		description = item.description;
 	}
 	return (
-		<div/>
+		<div
+			className={classNames(theme.item, {
+				[theme.disabled]: item.disabled,
+				[theme.selected]: value === title,
+			})}
+			title={title}
+			data-feature={item['data-feature'] || rest['data-feature']}
+		>
+			{get(item, 'icon') && <Icon className={theme['item-icon']} {...item.icon} />}
+			<div className={theme['item-text']}>
+				<span className={classNames(theme['item-title'], 'tc-typeahead-item-title')}>
+					<Emphasis value={value} text={title} />
+				</span>
+				{description && (
+					<p className={classNames(theme['item-description'], 'tc-typeahead-item-description')}>
+						<Emphasis value={value} text={description} />
+					</p>
+				)}
+			</div>
+		</div>
 	);
 }