+++ packages/components/src/index.js
@@ -56,9 +56,11 @@ import ObjectViewer from './ObjectViewer';
 import OverlayTrigger from './OverlayTrigger';
 import PieChart from './PieChart';
 import Progress from './Progress';
+import QualityBar from './QualityBar';
 import RadarChart from './RadarChart';
 import ResourceList from './ResourceList';
 import ResourcePicker from './ResourcePicker';
+import RatioBar from './RatioBar';
 import Rich from './Rich';
 import SidePanel from './SidePanel';
 import Skeleton from './Skeleton';
@@ -131,7 +133,9 @@ export {
 	OverlayTrigger,
 	PieChart,
 	Progress,
+	QualityBar,
 	RadarChart,
+	RatioBar,
 	ResourceList,
 	ResourcePicker,
 	Rich,