+++ packages/faceted-search/CHANGELOG.md
@@ -28,6 +28,8 @@ Types of changes
 
 - [Added](https://github.com/Talend/ui/pull/2994): Read only support
 - [Added](https://github.com/Talend/ui/pull/2997): Handle badges without labels
+- [Added](https://github.com/Talend/ui/pull/3000): Added complies/wordComplies pattern operators support
+- [Added](https://github.com/Talend/ui/pull/3013): Added colored badge support
 
 ## [0.13.0]